pyproject.toml
[tool.poetry]
name = "drf-api-permissions"
version = "0.1.0"
description = ""
authors = ["rubakanaan <rubajkanaan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
